requirements.dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.dev.in
#
backports.entry-points-selectable==1.1.1
    # via virtualenv
cfgv==3.3.1
    # via pre-commit
distlib==0.3.3
    # via virtualenv
filelock==3.3.2
    # via virtualenv
identify==2.3.5
    # via pre-commit
mypy==0.910
    # via
    #   -r requirements.dev.in
    #   sqlalchemy2-stubs
mypy-extensions==0.4.3
    # via mypy
nodeenv==1.6.0
    # via pre-commit
platformdirs==2.4.0
    # via virtualenv
pre-commit==2.15.0
    # via -r requirements.dev.in
pyyaml==6.0
    # via pre-commit
six==1.16.0
    # via virtualenv
sqlalchemy2-stubs==0.0.2a4
    # via -r requirements.dev.in
toml==0.10.2
    # via
    #   mypy
    #   pre-commit
types-requests==2.26.0
    # via -r requirements.dev.in
typing-extensions==3.10.0.2
    # via
    #   mypy
    #   sqlalchemy2-stubs
virtualenv==20.10.0
    # via pre-commit
